+++ CHANGELOG.markdown
@@ -1,3 +1,7 @@
+next
+----
+* TH code no longer generates superfluous `mempty` or `pure` subexpressions in derived `Bifoldable` or `Bitraversable` instances, respectively
+
 5.2.1
 ----
 * Added `Bifoldable` and `Bitraversable` instances for `Constant` from `transformers`